move the net_packet event to the net package
